requirements.txt
django==2.0.7
factory_boy==2.11.1
django-widget-tweaks==1.4.2
djangorestframework==3.8.2